red-knot: move symbol lookups in `symbol.rs`
