+++ crates/red_knot_python_semantic/src/types/infer.rs
@@ -48,8 +48,10 @@ use crate::semantic_index::expression::{Expression, ExpressionKind};
 use crate::semantic_index::semantic_index;
 use crate::semantic_index::symbol::{NodeWithScopeKind, NodeWithScopeRef, ScopeId};
 use crate::semantic_index::SemanticIndex;
-use crate::stdlib::builtins_module_scope;
-use crate::symbol::LookupError;
+use crate::symbol::{
+    builtins_module_scope, builtins_symbol, symbol, symbol_from_bindings, symbol_from_declarations,
+    typing_extensions_symbol, LookupError,
+};
 use crate::types::call::{Argument, CallArguments};
 use crate::types::diagnostic::{
     report_invalid_arguments_to_annotated, report_invalid_assignment,
@@ -64,13 +66,12 @@ use crate::types::diagnostic::{
 use crate::types::mro::MroErrorKind;
 use crate::types::unpacker::{UnpackResult, Unpacker};
 use crate::types::{
-    builtins_symbol, symbol, symbol_from_bindings, symbol_from_declarations, todo_type,
-    typing_extensions_symbol, Boundness, CallDunderResult, Class, ClassLiteralType, DynamicType,
-    FunctionType, InstanceType, IntersectionBuilder, IntersectionType, IterationOutcome,
-    KnownClass, KnownFunction, KnownInstanceType, MetaclassCandidate, MetaclassErrorKind,
-    RequiresExplicitReExport, SliceLiteralType, SubclassOfType, Symbol, SymbolAndQualifiers,
-    Truthiness, TupleType, Type, TypeAliasType, TypeAndQualifiers, TypeArrayDisplay,
-    TypeQualifiers, TypeVarBoundOrConstraints, TypeVarInstance, UnionBuilder, UnionType,
+    todo_type, Boundness, CallDunderResult, Class, ClassLiteralType, DynamicType, FunctionType,
+    InstanceType, IntersectionBuilder, IntersectionType, IterationOutcome, KnownClass,
+    KnownFunction, KnownInstanceType, MetaclassCandidate, MetaclassErrorKind, SliceLiteralType,
+    SubclassOfType, Symbol, SymbolAndQualifiers, Truthiness, TupleType, Type, TypeAliasType,
+    TypeAndQualifiers, TypeArrayDisplay, TypeQualifiers, TypeVarBoundOrConstraints,
+    TypeVarInstance, UnionBuilder, UnionType,
 };
 use crate::unpack::Unpack;
 use crate::util::subscript::{PyIndex, PySlice};
@@ -872,25 +873,22 @@ impl<'db> TypeInferenceBuilder<'db> {
         let use_def = self.index.use_def_map(binding.file_scope(self.db()));
         let declarations = use_def.declarations_at_binding(binding);
         let mut bound_ty = ty;
-        let declared_ty =
-            symbol_from_declarations(self.db(), declarations, RequiresExplicitReExport::No)
-                .map(|SymbolAndQualifiers(s, _)| {
-                    s.ignore_possibly_unbound().unwrap_or(Type::unknown())
-                })
-                .unwrap_or_else(|(ty, conflicting)| {
-                    // TODO point out the conflicting declarations in the diagnostic?
-                    let symbol_table = self.index.symbol_table(binding.file_scope(self.db()));
-                    let symbol_name = symbol_table.symbol(binding.symbol(self.db())).name();
-                    self.context.report_lint(
-                        &CONFLICTING_DECLARATIONS,
-                        node,
-                        format_args!(
-                            "Conflicting declared types for `{symbol_name}`: {}",
-                            conflicting.display(self.db())
-                        ),
-                    );
-                    ty.inner_type()
-                });
+        let declared_ty = symbol_from_declarations(self.db(), declarations)
+            .map(|SymbolAndQualifiers(s, _)| s.ignore_possibly_unbound().unwrap_or(Type::unknown()))
+            .unwrap_or_else(|(ty, conflicting)| {
+                // TODO point out the conflicting declarations in the diagnostic?
+                let symbol_table = self.index.symbol_table(binding.file_scope(self.db()));
+                let symbol_name = symbol_table.symbol(binding.symbol(self.db())).name();
+                self.context.report_lint(
+                    &CONFLICTING_DECLARATIONS,
+                    node,
+                    format_args!(
+                        "Conflicting declared types for `{symbol_name}`: {}",
+                        conflicting.display(self.db())
+                    ),
+                );
+                ty.inner_type()
+            });
         if !bound_ty.is_assignable_to(self.db(), declared_ty) {
             report_invalid_assignment(&self.context, node, declared_ty, bound_ty);
             // allow declarations to override inference in case of invalid assignment
@@ -910,10 +908,9 @@ impl<'db> TypeInferenceBuilder<'db> {
         let use_def = self.index.use_def_map(declaration.file_scope(self.db()));
         let prior_bindings = use_def.bindings_at_declaration(declaration);
         // unbound_ty is Never because for this check we don't care about unbound
-        let inferred_ty =
-            symbol_from_bindings(self.db(), prior_bindings, RequiresExplicitReExport::No)
-                .ignore_possibly_unbound()
-                .unwrap_or(Type::Never);
+        let inferred_ty = symbol_from_bindings(self.db(), prior_bindings)
+            .ignore_possibly_unbound()
+            .unwrap_or(Type::Never);
         let ty = if inferred_ty.is_assignable_to(self.db(), ty.inner_type()) {
             ty
         } else {
@@ -3312,11 +3309,7 @@ impl<'db> TypeInferenceBuilder<'db> {
         // If we're inferring types of deferred expressions, always treat them as public symbols
         let local_scope_symbol = if self.is_deferred() {
             if let Some(symbol_id) = symbol_table.symbol_id_by_name(symbol_name) {
-                symbol_from_bindings(
-                    db,
-                    use_def.public_bindings(symbol_id),
-                    RequiresExplicitReExport::No,
-                )
+                symbol_from_bindings(db, use_def.public_bindings(symbol_id))
             } else {
                 assert!(
                     self.deferred_state.in_string_annotation(),
@@ -3326,11 +3319,7 @@ impl<'db> TypeInferenceBuilder<'db> {
             }
         } else {
             let use_id = name_node.scoped_use_id(db, scope);
-            symbol_from_bindings(
-                db,
-                use_def.bindings_at_use(use_id),
-                RequiresExplicitReExport::No,
-            )
+            symbol_from_bindings(db, use_def.bindings_at_use(use_id))
         };
 
         let symbol = local_scope_symbol.or_fall_back_to(db, || {
